Support for SQL aggregates
