+++ sql/src/main/java/com/vmware/ddlog/translator/TranslationContext.java
@@ -31,12 +31,15 @@ class TranslationContext {
      */
     private boolean searchScopeName;
     private final List<Scope> translationScope;
-
-    private final HashMap<String, DDlogEBinOp.BOp> arithmeticFunctions;
-    private final HashMap<String, DDlogEBinOp.BOp> stringFunctions;
-    private final HashMap<String, DDlogEBinOp.BOp> booleanFunctions;
+    /**
+     * If a Node (representing a SQL expression) is mapped to an expression
+     * then the expression translation will return directly the expression
+     * instead of performing the translation proper.
+     */
+    private final HashMap<Node, DDlogExpression> substitutions;
 
     TranslationContext() {
+        this.substitutions = new HashMap<Node, DDlogExpression>();
         this.program = new DDlogProgram();
         this.program.imports.add(new DDlogImport("sql", ""));
         this.program.imports.add(new DDlogImport("sqlop", ""));
@@ -44,155 +47,10 @@ class TranslationContext {
         this.etv = new ExpressionTranslationVisitor();
         this.translationScope = new ArrayList<Scope>();
         this.searchScopeName = false;
-        this.arithmeticFunctions = new HashMap<String, DDlogEBinOp.BOp>();
-        this.stringFunctions = new HashMap<String, DDlogEBinOp.BOp>();
-        this.booleanFunctions = new HashMap<String, DDlogEBinOp.BOp>();
-        this.initializeMaps();
     }
 
     public DDlogExpression operationCall(DDlogEBinOp.BOp op, DDlogExpression left, DDlogExpression right) {
-        return this.etv.operationCall(op, left, right, this);
-    }
-
-    String getFunction(DDlogEBinOp.BOp op, DDlogType ltype, @Nullable DDlogType rtype) {
-        HashMap<String, DDlogEBinOp.BOp> map;
-        if (ltype.as(DDlogTBool.class) != null) {
-            map = this.booleanFunctions;
-        } else if (ltype.as(DDlogTInt.class) != null ||
-            ltype.as(DDlogTBit.class) != null ||
-            ltype.as(DDlogTSigned.class) != null) {
-            map = this.arithmeticFunctions;
-        } else {
-            map = this.stringFunctions;
-        }
-
-        String suffixl = ltype.mayBeNull ? "N" : "R";
-        String suffixr = rtype == null ? "" : (rtype.mayBeNull ? "N" : "R");
-        for (String k: map.keySet()) {
-            if (map.get(k).equals(op)) {
-                return k + "_" + suffixl + suffixr;
-            }
-        }
-        throw new RuntimeException("Could not find `" + op + "` for type " + ltype.toString());
-    }
-
-    private void initializeMaps() {
-        this.arithmeticFunctions.put("a_eq", DDlogEBinOp.BOp.Eq);
-        this.arithmeticFunctions.put("a_neq", DDlogEBinOp.BOp.Neq);
-        this.arithmeticFunctions.put("a_lt", DDlogEBinOp.BOp.Lt);
-        this.arithmeticFunctions.put("a_gt", DDlogEBinOp.BOp.Gt);
-        this.arithmeticFunctions.put("a_lte", DDlogEBinOp.BOp.Lte);
-        this.arithmeticFunctions.put("a_gte", DDlogEBinOp.BOp.Gte);
-        this.arithmeticFunctions.put("a_plus", DDlogEBinOp.BOp.Plus);
-        this.arithmeticFunctions.put("a_minus", DDlogEBinOp.BOp.Minus);
-        this.arithmeticFunctions.put("a_mod", DDlogEBinOp.BOp.Mod);
-        this.arithmeticFunctions.put("a_times", DDlogEBinOp.BOp.Times);
-        this.arithmeticFunctions.put("a_div", DDlogEBinOp.BOp.Div);
-        // The following don't work for 64 bits
-        // this.arithmeticFunctions.put("a_shiftr", DDlogEBinOp.BOp.ShiftR);
-        // this.arithmeticFunctions.put("a_shiftl", DDlogEBinOp.BOp.ShiftL);
-        this.arithmeticFunctions.put("a_band", DDlogEBinOp.BOp.BAnd);
-        this.arithmeticFunctions.put("a_bor", DDlogEBinOp.BOp.BOr);
-        this.arithmeticFunctions.put("a_bxor", DDlogEBinOp.BOp.BXor);
-
-        this.stringFunctions.put("s_concat", DDlogEBinOp.BOp.Concat);
-        this.stringFunctions.put("s_eq", DDlogEBinOp.BOp.Eq);
-        this.stringFunctions.put("s_neq", DDlogEBinOp.BOp.Neq);
-
-        this.booleanFunctions.put("b_eq", DDlogEBinOp.BOp.Eq);
-        this.booleanFunctions.put("b_neq", DDlogEBinOp.BOp.Neq);
-        this.booleanFunctions.put("b_and", DDlogEBinOp.BOp.And);
-        this.booleanFunctions.put("b_or", DDlogEBinOp.BOp.Or);
-        this.booleanFunctions.put("b_impl", DDlogEBinOp.BOp.Impl);
-    }
-
-    /**
-     * Generates a library with many functions that operate on Option values.
-     */
-    DDlogProgram generateLibrary() {
-        DDlogProgram result = new DDlogProgram();
-        for (HashMap<String, DDlogEBinOp.BOp> h: Arrays.asList(
-            this.arithmeticFunctions, this.booleanFunctions, this.stringFunctions)) {
-            for (String f : h.keySet()) {
-                DDlogEBinOp.BOp op = h.get(f);
-                for (int i = 0; i < 4; i++) {
-                    String function = f + "_";
-                    DDlogType leftType;
-                    DDlogType rightType;
-                    DDlogType raw;
-                    DDlogType withNull;
-                    if (h.equals(this.stringFunctions)) {
-                        raw = DDlogTString.instance;
-                    } else if (op.isBoolean()) {
-                        raw = DDlogTBool.instance;
-                    } else {
-                        raw = new DDlogTSigned(64, false);
-                    }
-                    withNull = raw.setMayBeNull(true);
-                    DDlogExpression leftMatch = new DDlogEVarDecl("l", raw);
-                    DDlogExpression rightMatch = new DDlogEVarDecl("r", raw);
-                    if ((i & 1) == 1) {
-                        function += "N";
-                        leftType = withNull;
-                        leftMatch = new DDlogEStruct("Some", Collections.singletonList(
-                            new DDlogEStruct.FieldValue("x",
-                                leftMatch)), leftType);
-                    } else {
-                        function += "R";
-                        leftType = raw;
-                    }
-                    if ((i & 2) == 2) {
-                        function += "N";
-                        rightType = withNull;
-                        rightMatch = new DDlogEStruct("Some", Collections.singletonList(
-                            new DDlogEStruct.FieldValue("x", rightMatch)), rightType);
-                    } else {
-                        function += "R";
-                        rightType = raw;
-                    }
-                    /*
-                    function add(left: Option<bigint>, right: Option<bigint>): Option<bigint> =
-                    match ((left, right)) {
-                        (Some{a}, Some{b}) -> Some{a + b},
-                        (_, _)             -> None
-                    }
-                    */
-                    DDlogFuncArg left = new DDlogFuncArg("left", false, leftType);
-                    DDlogFuncArg right = new DDlogFuncArg("right", false, rightType);
-                    DDlogType type = DDlogType.reduceType(leftType, rightType);
-                    if (op.isComparison()) {
-                        type = DDlogTBool.instance.setMayBeNull(type.mayBeNull);
-                    }
-                    DDlogExpression def;
-                    if (i == 0) {
-                        def = new DDlogEBinOp(op,
-                            new DDlogEVar("left", raw), new DDlogEVar("right", raw));
-                    } else {
-                        def = new DDlogEMatch(
-                            new DDlogETuple(
-                                new DDlogEVar("left", leftType),
-                                new DDlogEVar("right", rightType)),
-                            Arrays.asList(new DDlogEMatch.Case(
-                                    new DDlogETuple(leftMatch, rightMatch),
-                                    new DDlogEStruct("Some", Collections.singletonList(
-                                        new DDlogEStruct.FieldValue("x",
-                                            new DDlogEBinOp(op,
-                                                new DDlogEVar("l", raw), new DDlogEVar("r", raw)))
-                                    ), type)),
-                                new DDlogEMatch.Case(
-                                    new DDlogETuple(
-                                        new DDlogEPHolder(),
-                                        new DDlogEPHolder()),
-                                    new DDlogEStruct("None", Collections.emptyList(), type)))
-                        );
-                    }
-                    DDlogFunction func = new DDlogFunction(
-                        function, Arrays.asList(left, right), type, def);
-                    result.functions.add(func);
-                }
-            }
-        }
-        return result;
+        return this.etv.operationCall(op, left, right);
     }
 
     static String location(Node node) {
@@ -200,6 +58,23 @@ class TranslationContext {
             return "";
         NodeLocation location = node.getLocation().get();
         return "line: " + location.getLineNumber() + " column: " + location.getColumnNumber();
+    }
+
+    public void addSubstitution(Node node, DDlogExpression expression) {
+        this.substitutions.put(node, expression);
+    }
+
+    public void removeSubstitution(Node node) {
+        this.substitutions.remove(node);
+    }
+
+    public void clearSubstitutions() {
+        this.substitutions.clear();
+    }
+
+    @Nullable
+    public DDlogExpression getSubstitution(Node node) {
+        return this.substitutions.get(node);
     }
 
     void warning(String message, Node node) {
@@ -264,6 +139,8 @@ class TranslationContext {
     void exitScope() {
         this.translationScope.remove(this.translationScope.size() - 1);
     }
+
+    public Iterable<Scope> allScopes() { return this.translationScope; }
 
     DDlogExpression translateExpression(Expression expr) {
         return this.etv.process(expr, this);
