+++ sql/src/main/java/com/vmware/ddlog/util/Ternary.java
@@ -1,0 +1,7 @@
+package com.vmware.ddlog.util;
+
+public enum Ternary {
+    Yes,
+    No,
+    Maybe
+}
